clang tidy/format
