+++ RecoPixelVertexing/PixelTriplets/plugins/CAHitNtupletGeneratorKernelsAlloc.cu
@@ -1,1 +1,1 @@
-#include "CAHitNtupletGeneratorKernelsAlloc.h"
+#include "CAHitNtupletGeneratorKernelsAlloc.cc"
