+++ TrackingTools/AnalyticalJacobians/src/AnalyticalCurvilinearJacobianSSE.icc
@@ -1,9 +1,9 @@
-
-inline void AnalyticalCurvilinearJacobian::computeFullJacobian(const GlobalTrajectoryParameters& globalParameters,
-                                                               const GlobalPoint& x,
-                                                               const GlobalVector& p,
-                                                               const GlobalVector& h,
-                                                               const double& s) {
+// NOLINTNEXTLINE(misc-definitions-in-headers)
+void AnalyticalCurvilinearJacobian::computeFullJacobian(const GlobalTrajectoryParameters& globalParameters,
+                                                        const GlobalPoint& x,
+                                                        const GlobalVector& p,
+                                                        const GlobalVector& h,
+                                                        const double& s) {
   using mathSSE::Vec2D;
   using mathSSE::Vec3D;
   using mathSSE::Vec4D;
